backend/weather/requirements.txt
arrow==1.1.0
asgiref==3.2.10
blessed==1.18.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.5.0
django-picklefield==3.0.1
django-q==1.3.3
django-q-email==5.0.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
geojson==2.5.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyowm==3.2.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
